requirements.txt
streamlit
openai
langchain
sqlalchemy
pyodbc
pandas
numpy
python-dotenv
